+++ All.ui/miniprogram/pages/UI-Pages/ui.colorPicker/ui.colorPicker.js
@@ -28,6 +28,11 @@ Page({
     })
   },
 
+  
+  open: function() {
+    this.selectComponent('#colorPicker').open()
+  },
+
   /**
    * 生命周期函数--监听页面初次渲染完成
    */